Noticed that, in updateProgramName(), the code that starts the timer sequence for sending NRPNs out to the Mopho was inside the for-loop where the name character parameters are updated. This did not seem right, but when I moved the timer start out of the loop, the first name character would frequently not get updated. I tried starting nameCharCounter from -1 so there will be a delay before the first character gets updated and it seems to take care of the problem (though I'm not quite sure why).
Started moving code for clearing sequencer tracks into the processor. Also fiddled with the code for changing the program name. Instead of using a switch linked to a step counter in timerCallback(), I changed over to using if statements because I think that will make the code more compact when dealing with clearing sequencer tracks. Realized that, when you call setValueNotifyingHost() on a parameter, the parameter is flagged as changed even if its value hasn't actually changed. Therefore, NRPN messages always get sent out for all the characters when the program name is changed, and I don't need to split up updating the parameters and sending out the NRPNs to the Mopho.