Created gui_Logo.h and .cpp. Added path data for the Matrix-6G logo. Added an instance of the logo to the editor.
Mopho: In gui_Logo.cpp, renamed 'div_Logo_w' to 'logo_w' and renamed 'div_Logo_h' to 'logo_h'. Got rid of the intermediate path with the affine transform because it's left over from when I was treating the logo and its surrounding area as a "section" of the editor and there's no good reason for it now. Updated the position of the logo in the editor.